New approach to print automatically
Use of pyautogui resulted in more and more scenarios and problems to handle. After A LOT of research and testing I may have found a viable solution, which may actually work with automation in the task scheduler. Had to create a PDF file with "Lorem ipsum" text for the PGBK color and an extra image for all the other colors. This python script uses GhostScript to print the PDF file.